migrate file to imported
